chore(server): refacto process-effectifs
wip


wip


wip tests


wip tests





wip:tests


wip


wip


wip


wip
